% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_coords}
\alias{add_coords}
\alias{.add_coords}
\title{Add coordinates to an \code{sf} object.}
\usage{
.add_coords(point_sf, coord_names = c("longitude", "latitude"))
}
\arguments{
\item{point_sf}{A sf object.}

\item{coord_names}{A character. The names of the new columns.}
}
\value{
\preformatted{       A sf object.
}
}
\description{
This function adds coordinate columns to a sf object.
}
\author{
Alber Sánchez, \email{alber.ipia@inpe.br}
}
