% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deforestation_grid}
\alias{deforestation_grid}
\title{Deforestation grid for the Brazilian Amazon}
\format{
A data frame with 6783 rows and 2 variables:
\describe{
\item{id}{Cell identifier}
\item{geometry}{Cell geometry}
\item{area_km2}{Area of each cell in square kilometers}
}
}
\usage{
deforestation_grid
}
\description{
A dataset containing spatial data regarding the Brazilian Amazon.
These data represents a regular grid where each cell has an approximated
}
\keyword{datasets}
