% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.build_formula}
\alias{.build_formula}
\title{Build models for the given year.}
\usage{
.build_formula()
}
\value{
\preformatted{   A formula object.
}
}
\description{
This function builds formula objects.
}
