% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.get_model}
\alias{.get_model}
\title{Set the model.}
\usage{
.get_model(trees = 2000)
}
\arguments{
\item{trees}{An integer. The number of trees.}
}
\value{
A random forest model.
}
\description{
This function sets the parameters of the model used.
}
