% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.get_recipe}
\alias{.get_recipe}
\title{Set the recipe to be used.}
\usage{
.get_recipe(data_tb, my_formula)
}
\arguments{
\item{data_tb}{A tibble.}

\item{my_formula}{A formula object.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A recipe object.
}\if{html}{\out{</div>}}
}
\description{
This function sets the formula to be used by the classification model.
}
