% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.read_data}
\alias{.read_data}
\title{Read and prepare the data for the experiment.}
\usage{
.read_data(raw = FALSE)
}
\arguments{
\item{raw}{A logical. When TRUE, the whole data set is returned. Otherwise,
only those registers where deforestation is greater than zero are
returned.}
}
\value{
A tibble.
}
\description{
This functions read the data in the package (\code{deforestation_grid}), then
it transforms the deforested area from km2 to mt2, it applies a \code{log}
transformation, and filters out the data with no deforestation.
}
