% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{estimate_accuracy}
\alias{estimate_accuracy}
\title{Run a experiment several times and compute its accuracy.}
\usage{
estimate_accuracy(out_dir, iterations = 100)
}
\arguments{
\item{out_dir}{A path to directory for storing results.}

\item{iterations}{The number of times the experiment must be repeated.}
}
\value{
\preformatted{      A tibble
}
}
\description{
This functions fits a Random Forest model several times then estimates the
accuracy of the iterations of the model.
}
