% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{fit_model}
\alias{fit_model}
\title{This function fits a Random Forest model to the data in the package.}
\usage{
fit_model(out_dir)
}
\arguments{
\item{out_dir}{A path to directory for storing results.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     A tibble
}\if{html}{\out{</div>}}
}
\description{
This function fits a Random Forest model to the data in the package.
}
