% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{has_any_value}
\alias{has_any_value}
\alias{.has_any_value}
\title{Check if vector has a value.}
\usage{
.has_any_value(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A logical.
}
\description{
This checks if a vector has an element different from NA. It is
meant to be used with \code{dplyr::select_if} to select columns with data.
}
\author{
Alber Sánchez, \email{alber.ipia@inpe.br}
}
