% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{results_to_shp}
\alias{results_to_shp}
\title{This function joins the fitted model to the grid. It also saves the results
as a shapefile in the out directory under the name \code{priority_classes.shp}.}
\usage{
results_to_shp(out_dir)
}
\arguments{
\item{out_dir}{A path to directory for storing results.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     An sf object with the priority classes.
}\if{html}{\out{</div>}}
}
\description{
This function joins the fitted model to the grid. It also saves the results
as a shapefile in the out directory under the name \code{priority_classes.shp}.
}
