% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deforestation_grid}
\alias{deforestation_grid}
\title{Deforestation grid for the Brazilian Amazon}
\format{
A data frame with 6783 rows and 2 variables:
\describe{
\item{id}{Cell identifier}
\item{geometry}{Cell geometry}
\item{area_km2}{Area of each cell in square kilometers}
\item{agb_2018_forest_2021}{Biomass in Megagrams per hectare for 2018 after using the 2021 forest mask}
}
}
\usage{
deforestation_grid
}
\description{
A dataset containing spatial data regarding the Brazilian Amazon. These data
represents a regular grid where each cell has an approximated area of 25*25
kilometres.
}
\details{
The above ground biomass for 2018 was estimated from ESA Biomass Climate
Change Initiative (Biomass_cci) v3 as the mean of the pixels falling in each
cell after masking it using the areas identified as forest for 2021 by
PRODES.
}
\keyword{datasets}
