% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deforestation_grid}
\alias{deforestation_grid}
\title{Deforestation data for the Brazilian Amazon}
\format{
A data frame with 27132 rows and 14 variables:
\describe{
\item{id}{Cell identifier}
\item{area_PA}{area protected or indigenous lands in the grid (square kilometres)}
\item{dist_hidro}{distance to the closest waterway (kilometres)}
\item{dist_road}{distance to the closest highway (kilometres)}
\item{dist_road_hidro}{distance to the closest highway or to waterway (kilometres)}
\item{ref_year}{reference year}
\item{def}{deforestation in the reference year (square kilometres)}
\item{def_1_ly}{deforestation in the year before the reference year}
\item{def_2_ly}{deforestation in the two years before the reference year}
\item{def_4_ly}{deforestation in the four years before the reference year}
\item{dist_1_percent_ly}{distance to the closest grid centroid with more than 1\% deforestation in year before the reference year (kilometres)}
\item{dist_2_percent_ly}{distance to the closest grid centroid with more than 2\% deforestation in year before the reference year (kilometres)}
\item{active_fires_ly}{Number of active fires in the year before the reference year}
}
}
\usage{
deforestation_grid
}
\description{
A dataset containing data regarding deforestation in the Brazilian Amazon.
The data have been aggregated into a regular grid made of cell of of 25
square kilometres.
}
\keyword{datasets}
