% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.get_workflow}
\alias{.get_workflow}
\title{Set the workflow.}
\usage{
.get_workflow(recipe, spec)
}
\arguments{
\item{recipe}{A recipe object.}

\item{spec}{A model specification.}
}
\value{
\preformatted{   A workflow object.
}
}
\description{
This functions set the workflow of the model.
}
