% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.match_grid_year}
\alias{.match_grid_year}
\title{Join results to deforestation grid.}
\usage{
.match_grid_year(y, results_tb)
}
\arguments{
\item{y}{A character representing a single year.}

\item{results_tb}{A tibble with the results of fitting the model.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      An sf object.
}\if{html}{\out{</div>}}
}
\description{
The utility function takes the results, filters them by one year, classifies
them into Low, Average, and High using the 0.7 and 0.9 quantiles, and joins
the data to the deforestation grid.
}
