% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.read_data}
\alias{.read_data}
\title{Read and prepare the data for the experiment.}
\usage{
.read_data(raw = FALSE)
}
\arguments{
\item{raw}{A logical. When TRUE, the whole data set is returned. Otherwise,
only those registers where deforestation is greater than zero are
returned.}
}
\value{
A tibble.
}
\description{
This function reads the deforestation data from the package
(\code{deforestation_data}), then it transforms their areas from km2 to mt2, it
applies a \code{log} transformation, and finally, it filters out the data with no
deforestation.
}
