% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{.tune_model}
\alias{.tune_model}
\title{Tune model}
\usage{
.tune_model(workflow, folds, grid = 11)
}
\arguments{
\item{workflow}{A workflow object.}

\item{folds}{A fold object.}

\item{grid}{An integer.}
}
\value{
\preformatted{    A tune_grid object.
}
}
\description{
This functions tunes the model parameters
}
