% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{estimate_accuracy}
\alias{estimate_accuracy}
\title{Run a experiment several times and compute its accuracy.}
\usage{
estimate_accuracy(out_dir, iterations = 100)
}
\arguments{
\item{out_dir}{A path to directory for storing results.}

\item{iterations}{The number of times the experiment must be repeated.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A tibble
}\if{html}{\out{</div>}}
}
\description{
This function fits several Random Forest models to the data and then
estimates their accuracy.
}
\examples{

\dontrun{
library(prioritizedeforestationhotspots)
out_dir <- "~/Documents/prioritize_res"
estimate_accuracy(out_dir)
}

}
