% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{fit_model}
\alias{fit_model}
\title{Fits a model to the data in the package.}
\usage{
fit_model(out_dir)
}
\arguments{
\item{out_dir}{A path to directory for storing results.}
}
\value{
\preformatted{     A tibble
}
}
\description{
Fits a model to the data in the package.
}
