% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritize.R
\name{fit_model}
\alias{fit_model}
\title{This function fits a Random Forest model to the data in the package.}
\usage{
fit_model(out_dir, seed = 42)
}
\arguments{
\item{out_dir}{A path to directory for storing results.}

\item{seed}{An integer. A seed number to pass to \code{set.seed}.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     A tibble
}\if{html}{\out{</div>}}
}
\description{
This function fits a Random Forest model to the data in the package.
}
\examples{
\dontrun{
library(sf)
library(prioritizedeforestationhotspots)
out_dir <- "~/Documents/prioritize_res"
fit_model(out_dir)
}

}
